refactor: tests/integration/commands/addons/addons.test.mjs (#5887)
- convert to ESM
- replace ava with vitest
